Create experience_on_meetup_event
The event that I attained was set up by code for dc, a non-profit organization which uses technology such as data and algorithm to solve problems in the community. For each meeting, there will be several projects that require a variety of skill set such as developer, coding, data-analyst, etc.Basically, this meeting is to find a volunteer to meet the right project. After spending two hours in the meeting, I couldn't find any project that I can participate. All the project requires the skill set that I do not have any experience, even though some project needs someone who can analyze data, but I feel less competent than another volunteer. The feeling that I get in the meeting is that I'm not ready to fly yet because of my coding experience, knowledge in the field, and confidence. Those reasons made me feel like I do not belong in the meeting. Anyway, I will use this sort of feeling to be the fuel of motivation in the future